% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gglabeledcontour-package.R
\docType{package}
\name{gglabledcontour}
\alias{gglabledcontour}
\alias{gglabledcontour-package}
\title{Contour w/ Labels}
\description{
Simple geom that extends geom_contour() with labels
}

